refactor(outlook-semantic-mcp): extract resolveDirectoryIds as shared pure utility
Pull sanitizeWrongDirectoryIds out of CleanupSearchConditionsForUserQuery
into a standalone resolve-directory-ids.util.ts so the logic can be
reused by the upcoming BuildMsGraphKqlBatchRequestsQuery.
